test feedlist_view rename for easier git change comparison
